Add menu item for running apps on the dGPU
gnome-shell has the ability to launch applications using the dGPU [1].
However, this requires using the open source drivers and depends on
vga_switcheroo. The proprietary drivers are not (yet) supported [2].

Add a menu item here to run with the proprietary Nvidia drivers. The
menu item is only present while in hybrid mode.

[1]: https://bugzilla.gnome.org/show_bug.cgi?id=773117
[2]: https://gitlab.gnome.org/GNOME/gnome-shell/issues/1810